mmc: Port syslog() type fix in mmc_logstats() from thttpd v2.28

+++ src/mmc.c
@@ -639,8 +639,8 @@ static unsigned int hash(ino_t ino, dev_t dev, off_t size, time_t ctime)
 /* Generate debugging statistics syslog message. */
 void mmc_logstats(long secs)
 {
-	syslog(LOG_INFO, "map cache - %d allocated, %d active (%ld bytes), %d free; hash size: %d; expire age: %ld",
-	       alloc_count, map_count, (long int)mapped_bytes, free_count, hash_size, expire_age);
+	syslog(LOG_INFO, "map cache - %d allocated, %d active (%lld bytes), %d free; hash size: %d; expire age: %lld",
+	       alloc_count, map_count, (long long)mapped_bytes, free_count, hash_size, (long long)expire_age);
 
 	if (map_count + free_count != alloc_count)
 		syslog(LOG_ERR, "map counts don't add up!");
